Init. version of the rtl_fm launcher, internally store parameters as strings
The logic behind storing parameters as stringis is that it allows for pointers to the c_str() of each parameter
to be stored in array which is provided to execv(). In addition, the "option" variable for each parameter has
been converted to a string rather than a character so that "-<char for that cmd line option>" can be passed
to execv() through a pointer.

As of this version, exception handling needs to be done for the launcher.