Fix the segfault resulting from adding client update functionality
The problem was that each thread was being given a reference to a SocketWrapper instance inside the socketWrappersInUse vector.
Each time a new client would connect, this vector was resized, and all previous pointers to it were invalidated and their
use was undefined. Specifically, the shared_pointer to the socket was nullified, and accesses to it led to a dereference of zero (oops!)
I resolved this by using a std::list rather than a vector, since adding items doesn't result in previous
pointers to elements becoming invalidated (since an insert will only change a prev or next ptr). Using std::list meant
that SocketWrapper also needed to have == be overloaded, which I implemented as well.