Issue UPDATE_AVAILABLE upon volume changes
This commit allows changes the volume using the VOLUME= command to issue
an UPDATE_AVAILABLE response to all clients connected. Since VOLUME
doesn't get stored in the userProvidedCommands list stored in the
parameter  builder, this was somewhat of a hack. Specifically,
MY_STORED_CMDS doesn't return the volume, but CMDS_IN_USE does.
The reason for this weird hackyness is to cut down on client
side handling. If they see that UPDATE_AVAILABLE has been received,
they can call CMDS_IN_USE to get any and all commands which are
running on the server (be it parameters to rtl_fm or the volume).
The alternative was to have a special VOLUME_UPDATE_AVAILABLE
string broadcast when the volume was changed, which the user
would have to manually parse and call an associated GET_VOLUME
command.

In addition, since VOLUME isn't part of the param builder, note
that the volume can be set at any time, and the current value
will always be reflected in CMDS_IN_USE.